Data validation

+++ api/Controllers/CommentController.cs
@@ -29,6 +29,9 @@ namespace api.Controllers
         [HttpGet]
         public async Task<IActionResult> GetAll()
         {
+            if(!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var comments =  await _commentRepo.GetAllAsync();
             var CommentDto = comments.Select(s => s.ToCommentDto());
 
@@ -36,8 +39,12 @@ namespace api.Controllers
         }
 
 
-        [HttpGet("{id}")]
+        [HttpGet("{id:int}")]
         public async Task<IActionResult> GetById([FromRoute] int id) {
+            
+            if(!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var comment = await _commentRepo.GetByIdAsync(id);
             if(comment == null) {
                 return NotFound();
@@ -46,9 +53,12 @@ namespace api.Controllers
         }
 
 
-        [HttpPost("{stockId}")]
+        [HttpPost("{stockId:int}")]
         public async Task<IActionResult> Create([FromRoute] int stockId, CreateCommentDto commentDto)
         {
+            if(!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             if(!await _stockRepo.StockExists(stockId))
             {
                 return BadRequest("Stock does not exist");
@@ -61,10 +71,13 @@ namespace api.Controllers
 
 
         // [HttpPut("{id}")]
-        [HttpPut]
-        [Route("{id}")]
+        [HttpPut("{id:int}")]
+        // [Route("{id: int}")]
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateCommentRequestDto updateDto)
         {
+            if(!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var comment = await _commentRepo.UpdateAsync(id, updateDto.ToCommenFromUpdate());
 
             if(comment == null)
@@ -77,10 +90,12 @@ namespace api.Controllers
 
 
         // [HttpDelete("{id}")]
-        [HttpDelete]
-        [Route("{id}")]
+        [HttpDelete("{id:int}")]
+        // [Route("{id: int}")]
         public async Task<IActionResult> Delete([FromRoute] int id)
         {
+            if(!ModelState.IsValid)
+                return BadRequest(ModelState);
             var commentModel = await _commentRepo.DeleteAsync(id);
 
             {
